fix: don't add a link if parentRef is already linked to ref
Certain email threads can have References in differing orders which can
create an issue when parsing. When parsing other emails within the tree,
it is possible for a node have it's parent set to an node which is
already the self-node's child. For example, B has parent A and also
child A. This situation creates an infinite loop in the findChild
routine, resulting in a runtime panic.

Per the JWZ algorithm:
	Do not add a link if adding that link would introduce a loop: that is,
	before asserting A->B, search down the children of B to see if A is
	reachable, and also search down the children of A to see if B is
	reachable. If either is already reachable as a child of the other, don't
	add the link.

The implementation currently only inspects for B in the children of A.
Add a check for A in the children of B before creating new links. Add
two test emails which create a panic when threading without this check.

Update test values to reflect new threading results. Note that 6 emails
in the original test suite are now included in the threads that weren't
before. The passing test quantities are therefore 6 + 2 = 8 additional
emails in the threaded view. See attached images for the before and
after views of the threads.