fix: Adjust malformed headers in test data, upgrade dependencies
As we rely on enmime to parse the headers of emails, and it is now much stricter (or rather
it reports fatal errors on malformed headers), the tests would fail.

Examination of the errors involved shows bugs in the original mail clients, or perhaps the
servers, or perhaps even the original dump. None of these mail clients and none of the servers
are in use in any significant way anymore.

Hence, it seems better to fix the test data, and require users to perform some preprocessing
of ancient archives, in order to bring them up to the standards that even existed at the
time. Some very poor clients and servers were written back in the days when I were a lad,
(but not by me).